Progress - February 12th, 2024.
February 12th, 2024. Today was a great day for debugging and deepening knowledge about filtering, loc, and the legendary 'SettingWithCopyWarning: A value is trying to be set on a copy of a slice from a DataFrame' that pops up when slicing through a view instead of the main DataFrame in pandas.

Added the new category column
Transformed the 'Transaction Date' column to datetime to sort dates in descending order, dtype was object previously.
Added additional keywords to make the category lists more robust. By including all expense emails from Scotiabank, many 'Uncategorized' rows came up.
Defined the function to categorize said expenses automatically.
Created 'current month' variable to filter expenses for the ongoing month.
Created a new column 'Transaction Month' to effectively group later.
Grouped and summed expenses for the current month by 'Category' and 'Transaction Month'.
Created a copy of the DataFrame 'current_month_expenses_df' to make sure changes were not done on the original to debug for 'SettingWithCopyWarning'.
Current output: 
<img width="1014" alt="image" src="https://github.com/Victor-Dona/budget_automation_tool/assets/158128371/f0b026b7-2248-45db-8cc3-c3b120a378c7">